ft: add support to @adonisjs/bouncer

+++ src/extensions/crud.ts
@@ -1,30 +1,40 @@
+import { Bouncer } from '@adonisjs/bouncer'
+import { HttpContext as HttpContextOriginal } from '@adonisjs/core/http'
 import app from '@adonisjs/core/services/app'
-import { HttpContext } from '@adonisjs/core/http'
 import { BaseModel } from '@adonisjs/lucid/orm'
 import { VineValidator } from '@vinejs/vine'
 import type { Infer, SchemaTypes } from '@vinejs/vine/types'
 
-import { success, error } from './http.js'
-import { OutputFormatFn } from './format.js'
+import { ExtractScopes } from '@adonisjs/lucid/types/model'
 import { DreamerConfig } from '../../types.js'
 import { pagination } from '../validator/pagination.js'
+import { OutputFormatFn } from './format.js'
+import { success } from './http.js'
 import { parseModelSeachableFields, Scopes, SearchableModel } from './lucid.js'
-import { ExtractScopes } from '@adonisjs/lucid/types/model'
 
 export type GenericMetadata = Record<string, any> | undefined
+
+type BoucerAbility = ReturnType<(typeof Bouncer)['ability']>
+
+interface HttpContext extends HttpContextOriginal {
+  bouncer: InstanceType<typeof Bouncer<any, any, any>>
+}
 
 export type ShowOptions<M extends typeof BaseModel> = {
   key?: string
   mutate?: (row: InstanceType<M>) => any
+  ability?: BoucerAbility
 }
 
 export type StoreOptions<M extends typeof BaseModel, Payload> = {
   mutate?: (row: InstanceType<M>, payload: Payload) => Promise<void> | void
+  ability?: BoucerAbility
 }
 
 export type UpdateOptions<M extends typeof BaseModel, Payload> = {
   key?: string
   mutate?: (row: InstanceType<M>, payload: Payload) => Promise<void> | void
+  ability?: BoucerAbility
 }
 
 const getIdKey = (key?: string) => {
@@ -43,75 +53,78 @@ export function index<
     perPage?: number
     formats?: OutputFormatFn<Model>[]
     scope?: Scopes<Model> | ((scopes: ExtractScopes<Model>) => void)
+    ability?: BoucerAbility
   }
 ) {
   return async (ctx: HttpContext) => {
-    const { request, response } = ctx
+    const { request, response, bouncer } = ctx
 
-    try {
-      const { validator = pagination, perPage = 10 } = options || {}
-      const { page = 1, ...params } = validator ? await request.validateUsing(validator as any) : {}
-      const baseQuery = model.query()
-      const searchableFields = parseModelSeachableFields(model)
-
-      for (const key in params) {
-        const operator = searchableFields[key] || 'equals'
-
-        if (operator === 'equals') {
-          baseQuery.where(key, params[key])
-        } else if (operator === 'like') {
-          baseQuery.where(key, 'LIKE', `%${params[key]}%`)
-        }
-      }
-
-      if (options?.scope) {
-        if (typeof options.scope === 'function') {
-          baseQuery.withScopes(options.scope)
-        }
-
-        if (typeof options.scope === 'string') {
-          baseQuery.withScopes((q: ExtractScopes<Model>) => {
-            q[options.scope]()
-          })
-        }
-      }
-
-      if (options?.formats) {
-        const { format, f } = ctx.request.qs()
-        const reqFormat = f || format
-
-        if (reqFormat) {
-          const formatFn = options.formats.find((c) => c.formatName === reqFormat)
-
-          if (formatFn) {
-            const rows = await baseQuery
-            return formatFn(ctx, rows)
-          }
-        }
-      }
-
-      const rows = await baseQuery.paginate(page, perPage)
-
-      return success(response, rows)
-    } catch (err) {
-      return error(response, err)
+    if (options?.ability) {
+      if (!bouncer) throw new Error('@adonisjs/bouncer is required to use abilities')
+      await bouncer.authorize(options.ability)
     }
+
+    const { validator = pagination, perPage = 10 } = options || {}
+    const { page = 1, ...params } = validator ? await request.validateUsing(validator as any) : {}
+    const baseQuery = model.query()
+    const searchableFields = parseModelSeachableFields(model)
+
+    for (const key in params) {
+      const operator = searchableFields[key] || 'equals'
+
+      if (operator === 'equals') {
+        baseQuery.where(key, params[key])
+      } else if (operator === 'like') {
+        baseQuery.where(key, 'LIKE', `%${params[key]}%`)
+      }
+    }
+
+    if (options?.scope) {
+      if (typeof options.scope === 'function') {
+        baseQuery.withScopes(options.scope)
+      }
+
+      if (typeof options.scope === 'string') {
+        baseQuery.withScopes((q: ExtractScopes<Model>) => {
+          q[options.scope]()
+        })
+      }
+    }
+
+    if (options?.formats) {
+      const { format, f } = ctx.request.qs()
+      const reqFormat = f || format
+
+      if (reqFormat) {
+        const formatFn = options.formats.find((c) => c.formatName === reqFormat)
+
+        if (formatFn) {
+          const rows = await baseQuery
+          return formatFn(ctx, rows)
+        }
+      }
+    }
+
+    const rows = await baseQuery.paginate(page, perPage)
+
+    return success(response, rows)
   }
 }
 
 export function show<Model extends typeof BaseModel>(model: Model, options?: ShowOptions<Model>) {
-  return async ({ params, response }: HttpContext) => {
-    try {
-      const row = await model.query().where(getIdKey(options?.key), params.id).firstOrFail()
+  return async ({ params, response, bouncer }: HttpContext) => {
+    const row = await model.query().where(getIdKey(options?.key), params.id).firstOrFail()
 
-      if (options?.mutate) {
-        return success(response, options.mutate(row))
-      }
-
-      return success(response, row.toJSON())
-    } catch (err) {
-      return error(response, err)
+    if (options?.ability) {
+      if (!bouncer) throw new Error('@adonisjs/bouncer is required to use abilities')
+      await bouncer.authorize(options.ability, row)
     }
+
+    if (options?.mutate) {
+      return success(response, options.mutate(row))
+    }
+
+    return success(response, row.toJSON())
   }
 }
 
@@ -124,22 +137,23 @@ export function store<
   validator: VineValidator<Schema, MetaData>,
   options?: StoreOptions<Model, Infer<Schema>>
 ) {
-  return async ({ request, response }: HttpContext) => {
-    try {
-      const data = await request.validateUsing(validator as any)
-      const row = new model()
+  return async ({ request, response, bouncer }: HttpContext) => {
+    const data = await request.validateUsing(validator as any)
+    const row = new model()
 
-      if (options?.mutate) {
-        await options.mutate(row as any, data as Infer<typeof validator>)
-      } else {
-        row.fill(data)
-      }
-
-      await row.save()
-      return success(response, row.toJSON(), 201)
-    } catch (err) {
-      return error(response, err)
+    if (options?.mutate) {
+      await options.mutate(row as any, data as Infer<typeof validator>)
+    } else {
+      row.fill(data)
     }
+
+    if (options?.ability) {
+      if (!bouncer) throw new Error('@adonisjs/bouncer is required to use abilities')
+      await bouncer.authorize(options.ability, row)
+    }
+
+    await row.save()
+    return success(response, row.toJSON(), 201)
   }
 }
 
@@ -152,35 +166,41 @@ export function update<
   validator: VineValidator<Schema, MetaData>,
   options?: UpdateOptions<Model, Infer<Schema>>
 ) {
-  return async ({ request, response, params }: HttpContext) => {
-    try {
-      const data = await request.validateUsing(validator as any)
-      const row = await model.query().where(getIdKey(options?.key), params.id).firstOrFail()
+  return async ({ request, response, params, bouncer }: HttpContext) => {
+    const data = await request.validateUsing(validator as any)
+    const row = await model.query().where(getIdKey(options?.key), params.id).firstOrFail()
 
-      if (options?.mutate) {
-        await options.mutate(row, data as Infer<typeof validator>)
-      } else {
-        row.merge(data)
-      }
-
-      await row.save()
-      return success(response, row.toJSON(), 201)
-    } catch (err) {
-      return error(response, err)
+    if (options?.ability) {
+      if (!bouncer) throw new Error('@adonisjs/bouncer is required to use abilities')
+      await bouncer.authorize(options.ability, row)
     }
+
+    if (options?.mutate) {
+      await options.mutate(row, data as Infer<typeof validator>)
+    } else {
+      row.merge(data)
+    }
+
+    await row.save()
+    return success(response, row.toJSON(), 201)
   }
 }
 
-export function destroy<Model extends typeof BaseModel>(model: Model) {
-  return async ({ params, response }: HttpContext) => {
-    try {
-      const row = await model.query().where(getIdKey(), params.id).firstOrFail()
-      await row.delete()
+export function destroy<Model extends typeof BaseModel>(
+  model: Model,
+  options?: { ability?: BoucerAbility }
+) {
+  return async ({ params, response, bouncer }: HttpContext) => {
+    const row = await model.query().where(getIdKey(), params.id).firstOrFail()
 
-      return success(response)
-    } catch (err) {
-      return error(response, err)
+    if (options?.ability) {
+      if (!bouncer) throw new Error('@adonisjs/bouncer is required to use abilities')
+      await bouncer.authorize(options.ability, row)
     }
+
+    await row.delete()
+
+    return success(response)
   }
 }
